types_tests: Rename to macros_tests
There are only tests of our derive macros in this file.
