+++ scylla-cql/src/macros_tests.rs
